improved buffer api

+++ kool-core/src/commonMain/kotlin/de/fabmax/kool/modules/audio/WavFile.kt
@@ -138,7 +138,7 @@ class WavFile(fileData: Uint8Buffer) {
         fun read(buf: ByteArray): Int {
             val len = buf.size.clamp(0, available())
             for (i in 0 until len) {
-                buf[i] = data[pos++]
+                buf[i] = data[pos++].toByte()
             }
             return len
         }
